% Generated by build script
\name{re3_repository}
\alias{re3_repository}
\title{Get detailed info about a research data repository}
\usage{
re3_repository()
}
\arguments{
\item{repo_id}{Repository ID from re3_search (e.g., "r3d100010468")}
}
\value{tibble: id, name, url, description, subjects, content_types,}
\description{Get detailed info about a research data repository}
